Resolve conflict and update e2e test case
